+++ alicloud/connectivity/client.go
@@ -2,7 +2,6 @@ package connectivity
 
 import (
 	"fmt"
-
 	"github.com/aliyun/alibaba-cloud-sdk-go/sdk"
 	"github.com/aliyun/alibaba-cloud-sdk-go/sdk/endpoints"
 	"github.com/aliyun/alibaba-cloud-sdk-go/sdk/requests"
@@ -20,6 +19,7 @@ import (
 	"github.com/aliyun/alibaba-cloud-sdk-go/services/ots"
 	"github.com/aliyun/alibaba-cloud-sdk-go/services/pvtz"
 	"github.com/aliyun/alibaba-cloud-sdk-go/services/r-kvstore"
+	"github.com/aliyun/alibaba-cloud-sdk-go/services/ram"
 	"github.com/aliyun/alibaba-cloud-sdk-go/services/rds"
 	"github.com/aliyun/alibaba-cloud-sdk-go/services/slb"
 	"github.com/aliyun/alibaba-cloud-sdk-go/services/sts"
@@ -34,7 +34,6 @@ import (
 	"github.com/denverdino/aliyungo/cs"
 	"github.com/denverdino/aliyungo/dns"
 	"github.com/denverdino/aliyungo/kms"
-	"github.com/denverdino/aliyungo/ram"
 	"github.com/dxh031/ali_mns"
 	"github.com/hashicorp/terraform/terraform"
 
@@ -68,7 +67,7 @@ type AliyunClient struct {
 	slbconn                      *slb.Client
 	ossconn                      *oss.Client
 	dnsconn                      *dns.Client
-	ramconn                      ram.RamClientInterface
+	ramconn                      *ram.Client
 	csconn                       *cs.Client
 	cdnconn                      *cdn.CdnClient
 	kmsconn                      *kms.Client
@@ -372,7 +371,7 @@ func (client *AliyunClient) WithDnsClient(do func(*dns.Client) (interface{}, err
 	return do(client.dnsconn)
 }
 
-func (client *AliyunClient) WithRamClient(do func(ram.RamClientInterface) (interface{}, error)) (interface{}, error) {
+func (client *AliyunClient) WithRamClient(do func(client *ram.Client) (interface{}, error)) (interface{}, error) {
 	goSdkMutex.Lock()
 	defer goSdkMutex.Unlock()
 
@@ -382,17 +381,17 @@ func (client *AliyunClient) WithRamClient(do func(ram.RamClientInterface) (inter
 		if endpoint == "" {
 			endpoint = loadEndpoint(client.config.RegionId, RAMCode)
 		}
-		if endpoint == "" {
-			endpoint = ram.RAMDefaultEndpoint
+		if strings.HasPrefix(endpoint, "http") {
+			endpoint = fmt.Sprintf("https://%s", strings.TrimPrefix(endpoint, "http://"))
 		}
-		if !strings.HasPrefix(endpoint, "http") {
-			endpoint = fmt.Sprintf("https://%s", strings.TrimPrefix(endpoint, "://"))
+		if endpoint != "" {
+			endpoints.AddEndpointMapping(client.config.RegionId, string(RAMCode), endpoint)
 		}
-		accessKey, secretKey, securityToken, err := client.config.getAuthCredentialByEcsRoleName()
+
+		ramconn, err := ram.NewClientWithOptions(client.config.RegionId, client.getSdkConfig(), client.config.getAuthCredential(true))
 		if err != nil {
-			return nil, err
+			return nil, fmt.Errorf("unable to initialize the RAM client: %#v", err)
 		}
-		ramconn := ram.NewClientWithEndpointAndSecurityToken(endpoint, accessKey, secretKey, securityToken)
 		client.ramconn = ramconn
 	}