+++ alicloud/resource_alicloud_ram_role_policy_attachment_test.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/denverdino/aliyungo/ram"
+	"github.com/aliyun/alibaba-cloud-sdk-go/services/ram"
 	"github.com/hashicorp/terraform/helper/acctest"
 	"github.com/hashicorp/terraform/helper/resource"
 	"github.com/hashicorp/terraform/terraform"
@@ -46,24 +46,23 @@ func testAccCheckRamRolePolicyAttachmentExists(n string, policy *ram.Policy, rol
 	return func(s *terraform.State) error {
 		rs, ok := s.RootModule().Resources[n]
 		if !ok {
-			return fmt.Errorf("Not found: %s", n)
+			return WrapError(fmt.Errorf("Not found: %s", n))
 		}
 
 		if rs.Primary.ID == "" {
-			return fmt.Errorf("No Attachment ID is set")
+			return WrapError(Error("No Attachment ID is set"))
 		}
 
 		client := testAccProvider.Meta().(*connectivity.AliyunClient)
 
-		request := ram.RoleQueryRequest{
-			RoleName: role.RoleName,
-		}
+		request := ram.CreateListPoliciesForRoleRequest()
+		request.RoleName = role.RoleName
 
-		raw, err := client.WithRamClient(func(ramClient ram.RamClientInterface) (interface{}, error) {
+		raw, err := client.WithRamClient(func(ramClient *ram.Client) (interface{}, error) {
 			return ramClient.ListPoliciesForRole(request)
 		})
 		if err == nil {
-			response, _ := raw.(ram.PolicyListResponse)
+			response, _ := raw.(*ram.ListPoliciesForRoleResponse)
 			if len(response.Policies.Policy) > 0 {
 				for _, v := range response.Policies.Policy {
 					if v.PolicyName == policy.PolicyName && v.PolicyType == policy.PolicyType {
@@ -71,9 +70,9 @@ func testAccCheckRamRolePolicyAttachmentExists(n string, policy *ram.Policy, rol
 					}
 				}
 			}
-			return fmt.Errorf("Error finding attach %s", rs.Primary.ID)
+			return WrapError(fmt.Errorf("Error finding attach %s", rs.Primary.ID))
 		}
-		return fmt.Errorf("Error finding attach %s: %#v", rs.Primary.ID, err)
+		return WrapError(err)
 	}
 }
 
@@ -87,22 +86,21 @@ func testAccCheckRamRolePolicyAttachmentDestroy(s *terraform.State) error {
 		// Try to find the attachment
 		client := testAccProvider.Meta().(*connectivity.AliyunClient)
 
-		request := ram.RoleQueryRequest{
-			RoleName: rs.Primary.Attributes["role_name"],
-		}
+		request := ram.CreateListPoliciesForRoleRequest()
+		request.RoleName = rs.Primary.Attributes["role_name"]
 
-		raw, err := client.WithRamClient(func(ramClient ram.RamClientInterface) (interface{}, error) {
+		raw, err := client.WithRamClient(func(ramClient *ram.Client) (interface{}, error) {
 			return ramClient.ListPoliciesForRole(request)
 		})
 
 		if err != nil && !RamEntityNotExist(err) {
-			return err
+			return WrapError(err)
 		}
-		response, _ := raw.(ram.PolicyListResponse)
+		response, _ := raw.(*ram.ListPoliciesForRoleResponse)
 		if len(response.Policies.Policy) > 0 {
 			for _, v := range response.Policies.Policy {
 				if v.PolicyName == rs.Primary.Attributes["policy_name"] && v.PolicyType == rs.Primary.Attributes["policy_type"] {
-					return fmt.Errorf("Error attachment still exist.")
+					return WrapError(Error("Error attachment still exist."))
 				}
 			}
 		}