+++ CHANGELOG.md
@@ -8,6 +8,7 @@ FEATURES:
 IMPROVEMENTS:
 
 - Improve api_gateway_vpc_access testcases [GH-738]
+- Modify errors and  ram client to dock with the GO SDK [GH-735]
 - provider supports getting credential via ecs role name [GH-731]
 - Update testcases for cen region domain route entries [GH-729]
 - cs_kubernetes supports user_ca [GH-726]